backend/requirements.txt
# --- Web framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# --- Database / ORM (Supabase Postgres via SQLAlchemy) ---
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0         # Postgres driver (works with Supabase)
alembic>=1.13.0                # migrations

# --- Pydantic (FastAPI uses it; pin explicitly to avoid surprises) ---
pydantic>=2.6.0

# --- Google / Gemini integration ---
google-genai>=0.1.0            # Gemini API client
google-adk>=0.1.0              # Google Agentic Dev Kit

# --- Opik observability / evaluation ---
opik>=0.1.0

# --- Config / environment ---
python-dotenv>=1.0.0           # load .env with Supabase URL, keys, etc.

# --- HTTP client (useful for any external APIs/tools) ---
httpx>=0.27.0

# --- LiteLLM for OpenAI/Anthropic  ---
litellm==1.77.1

# --- Faster JSON serialization (can be plugged into FastAPI) ---
orjson>=3.10.0

# --- PyJWT for token handling (if needed) ---
PyJWT>=2.8.0